+++ website/src/playground/components/Resizable.tsx
@@ -123,7 +123,7 @@ export default function Resizable({
 			window.removeEventListener("mouseup", onMouseUp);
 			window.removeEventListener("mousemove", onMouseMove);
 		};
-	});
+	}, [ref, isResizing, canResize, handler, sizeStore]);
 
 	useEffect(() => {
 		if (cursor === undefined) {